PIM-3595: re-organize savers
