+++ spec/Pim/Bundle/CatalogBundle/Doctrine/Common/Saver/ProductSaverSpec.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace spec\Pim\Bundle\CatalogBundle\Saver;
+namespace spec\Pim\Bundle\CatalogBundle\Doctrine\Common\Saver;
 
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Manager\CompletenessManager;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
-use Pim\Bundle\CatalogBundle\Saver\ProductSavingOptionsResolver;
+use Pim\Bundle\CatalogBundle\Doctrine\Common\Saver\ProductSavingOptionsResolver;
 use Prophecy\Argument;
 
 class ProductSaverSpec extends ObjectBehavior
