+++ src/Pim/Bundle/CatalogBundle/Doctrine/Common/Saver/ProductSavingOptionsResolver.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Saver;
+namespace Pim\Bundle\CatalogBundle\Doctrine\Common\Saver;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
